Refactor sort_widget to use custom SortType
Previously sort_widget used a boolean for every available
sort type, like ByName, ByDate, ByDuration.
This makes it hard to add more sort options and results in
overall less readable code.
To combat this, this commit adds an enum SortType, that is instead used
to express the type of sorting the sort widget should apply.