fix: system tray causing crash on linux
The
[system_tray](https://pub.dev/packages/system_tray)
package appears abandoned, with no update or
comments for 13 months. The issue this PR fixes
is due to a known bug in the system_tray package:
https://github.com/antler119/system_tray/issues/64

This PR replaces the system_tray package with the
[tray_manager](https://pub.dev/packages/tray_manager)
package, which is actively maintained and doesn't
cause the crash. `tray_manager` has beem around
for 3+ years and had a release just 11 days ago.

Resolves #232